Move R model view to its own module
