+++ superset/views/core.py
@@ -150,71 +150,6 @@ DATABASE_KEYS = [
 DATASOURCE_MISSING_ERR = __("The data source seems to have been deleted")
 
 
-class KV(BaseSupersetView):
-
-    """Used for storing and retrieving key value pairs"""
-
-    @event_logger.log_this
-    @has_access_api
-    @expose("/store/", methods=["POST"])
-    def store(self) -> FlaskResponse:
-        try:
-            value = request.form.get("data")
-            obj = models.KeyValue(value=value)
-            db.session.add(obj)
-            db.session.commit()
-        except Exception as ex:
-            return json_error_response(utils.error_msg_from_exception(ex))
-        return Response(json.dumps({"id": obj.id}), status=200)
-
-    @event_logger.log_this
-    @has_access_api
-    @expose("/<int:key_id>/", methods=["GET"])
-    def get_value(self, key_id: int) -> FlaskResponse:
-        try:
-            kv = db.session.query(models.KeyValue).filter_by(id=key_id).scalar()
-            if not kv:
-                return Response(status=404, content_type="text/plain")
-        except Exception as ex:
-            return json_error_response(utils.error_msg_from_exception(ex))
-        return Response(kv.value, status=200, content_type="text/plain")
-
-
-class R(BaseSupersetView):
-
-    """used for short urls"""
-
-    @event_logger.log_this
-    @expose("/<int:url_id>")
-    def index(self, url_id: int) -> FlaskResponse:
-        url = db.session.query(models.Url).get(url_id)
-        if url and url.url:
-            explore_url = "//superset/explore/?"
-            if url.url.startswith(explore_url):
-                explore_url += f"r={url_id}"
-                return redirect(explore_url[1:])
-            else:
-                return redirect(url.url[1:])
-        else:
-            flash("URL to nowhere...", "danger")
-            return redirect("/")
-
-    @event_logger.log_this
-    @has_access_api
-    @expose("/shortner/", methods=["POST"])
-    def shortner(self) -> FlaskResponse:
-        url = request.form.get("data")
-        obj = models.Url(url=url)
-        db.session.add(obj)
-        db.session.commit()
-        return Response(
-            "{scheme}://{request.headers[Host]}/r/{obj.id}".format(
-                scheme=request.scheme, request=request, obj=obj
-            ),
-            mimetype="text/plain",
-        )
-
-
 class Superset(BaseSupersetView):
     """The base views for Superset!"""
